feat: rename client to zeebe-client-spring
