+++ zeebe/clients/zeebe-client-spring/src/main/java/io/camunda/zeebe/spring/client/bean/BeanInfo.java
