+++ zeebe/clients/zeebe-client-spring/src/test/java/io/camunda/zeebe/spring/client/bean/ClassInfoTest.java
